melophonyApiServer/requirements.txt
django
pyjwt
youtube_dl